Just a commit

+++ app/src/main/java/com/driverapp/android/events/create/CreateActivity.java
@@ -8,7 +8,9 @@ import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
+import android.os.Environment;
 import android.provider.MediaStore;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -21,21 +23,24 @@ import android.widget.Toast;
 import com.driverapp.android.R;
 import com.driverapp.android.core.BaseActivity;
 import com.driverapp.android.core.Core;
+import com.driverapp.android.events.EventActivity;
 import com.driverapp.android.events.create.cropping.CropActivity;
 import com.driverapp.android.models.EventCategory;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 
 
 public class CreateActivity extends BaseActivity {
 
     private static final int PICK_IMAGE = 100;
+    private static final int PICK_IMAGE_NEW_API = 105;
     private static final int PICK_CATEGORY = 101;
     private static final int PICK_MAP = 102;
     private static final int PICK_PHOTO = 103;
     private static final int CROP_PICTURE = 104;
-    private ImageView imageView;
+    private TextView imagePickView;
     private EventCategory selectedCategory;
     private View pickCategoryView;
     private TextView pickCategoryTextView;
@@ -50,6 +55,8 @@ public class CreateActivity extends BaseActivity {
     private EditText bodyView;
     private Bitmap selectedImageBitmap;
     private File selectedImageFile;
+    private ImageView imagePickBackgroundView;
+    private View createButton;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -58,9 +65,17 @@ public class CreateActivity extends BaseActivity {
 
         bodyView = (EditText) findViewById(R.id.text);
 
-        pickCategoryView =  findViewById(R.id.pick_category);
+        pickCategoryView = findViewById(R.id.pick_category);
         pickCategoryTextView = (TextView) findViewById(R.id.pick_category_text);
         pickCategoryImageView = (ImageView) findViewById(R.id.pick_category_image);
+        createButton = findViewById(R.id.create);
+
+        createButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                create();
+            }
+        });
 
         pickCategoryView.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -75,41 +90,65 @@ public class CreateActivity extends BaseActivity {
         pickLocationView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                startActivityForResult(new Intent(getBaseContext(),LocationPickerActivity.class), PICK_MAP);
+                startActivityForResult(new Intent(getBaseContext(), LocationPickerActivity.class), PICK_MAP);
             }
         });
 
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        imageView = (ImageView) findViewById(R.id.image);
-        imageView.setOnClickListener(new View.OnClickListener() {
+        imagePickView = (TextView) findViewById(R.id.image_pick);
+        imagePickBackgroundView = (ImageView) findViewById(R.id.image_pick_background);
+        imagePickView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
 
                 new AlertDialog.Builder(CreateActivity.this)
                         .setTitle(R.string.create_pick_image_title)
                         .setItems(new String[]{getString(R.string.create_pick_image_gallery), getString(R.string.create_pick_image_photo)}, new DialogInterface.OnClickListener() {
-                            @Override
-                            public void onClick(DialogInterface dialog, int which) {
-                                if (which == 0) {
-                                    Intent getImageIntent = new Intent();
-                                    getImageIntent.setType("image/*");
-                                    getImageIntent.setAction(Intent.ACTION_GET_CONTENT);
-                                    if (getImageIntent.resolveActivity(getPackageManager()) != null) {
-                                        startActivityForResult(Intent.createChooser(getImageIntent, "Select Picture"), PICK_IMAGE);
-                                    }
-                                } else {
-                                    Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                                        if (takePictureIntent.resolveActivity(getPackageManager()) != null) {
-                                            startActivityForResult(takePictureIntent, PICK_PHOTO);
-                                        }
+                                    @Override
+                                    public void onClick(DialogInterface dialog, int which) {
+                                        if (which == 0) {
+                                            if (Build.VERSION.SDK_INT < 19) {
+                                                Intent getImageIntent = new Intent();
+                                                getImageIntent.setType("image/*");
+                                                getImageIntent.setAction(Intent.ACTION_GET_CONTENT);
+                                                if (getImageIntent.resolveActivity(getPackageManager()) != null) {
+                                                    startActivityForResult(Intent.createChooser(getImageIntent, "Select Picture"), PICK_IMAGE);
+                                                }
+                                            } else {
+                                                Intent intent = new Intent(Intent.ACTION_OPEN_DOCUMENT);
+                                                intent.addCategory(Intent.CATEGORY_OPENABLE);
+                                                intent.setType("image/jpeg");
+                                                startActivityForResult(intent, PICK_IMAGE_NEW_API);
+                                            }
+                                        } else {
+                                            Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+                                            if (takePictureIntent.resolveActivity(getPackageManager()) != null) {
+                                                selectedImageFile = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES)+"/DriverApp","picture"+ System.currentTimeMillis()/1000L+".jpg");
+                                                new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES)+"/DriverApp").mkdirs();
+                                                try {
+                                                    selectedImageFile.createNewFile();
+                                                } catch (IOException e) {
+                                                    e.printStackTrace();
+                                                }
+                                                String uripath = selectedImageFile.getPath();
+                                                takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, Uri.parse("file://"+uripath));
+                                                startActivityForResult(takePictureIntent, PICK_PHOTO);
+                                            }
 
+                                        }
+                                    }
                                 }
-                            }
-                        })
-                        .setCancelable(true)
-                        .show();
-            }
-        });
+
+                            )
+                                    .
+
+                            setCancelable(true)
+
+                            .
+
+                            show();
+                        }
+            });
     }
 
     @Override
@@ -118,7 +157,7 @@ public class CreateActivity extends BaseActivity {
         try {
             switch (requestCode) {
                 case CROP_PICTURE:
-                    if(resultCode == RESULT_OK) {
+                    if (resultCode == RESULT_OK) {
                         String path = data.getStringExtra(CropActivity.IMAGE_PATH_RESULT);
 
                         if (path == null) {
@@ -126,9 +165,43 @@ public class CreateActivity extends BaseActivity {
                         }
                         selectedImageFile = new File(path);
                         selectedImageBitmap = BitmapFactory.decodeFile(path);
-                        imageView.setImageBitmap(selectedImageBitmap);
+                        imagePickBackgroundView.setImageBitmap(selectedImageBitmap);
+                        imagePickView.setTextColor(0xffffffff);
                     }
                     break;
+                case PICK_IMAGE_NEW_API:
+                    if(resultCode==RESULT_OK) {
+                            Uri originalUri = data.getData();
+                            final int takeFlags = data.getFlags()
+                                    & (Intent.FLAG_GRANT_READ_URI_PERMISSION
+                                    | Intent.FLAG_GRANT_WRITE_URI_PERMISSION);
+                            // Check for the freshest data.
+                            getContentResolver().takePersistableUriPermission(originalUri, takeFlags);
+
+                            Bitmap capturedPhotoBitmap = null;
+
+                            capturedPhotoBitmap = BitmapFactory.decodeStream(getContentResolver().openInputStream(originalUri));
+                            File capturedPhotoFile = new File(getCacheDir(),
+                                    "temp"
+                            );
+                            if (capturedPhotoFile.exists()) {
+                                capturedPhotoFile.delete();
+                                capturedPhotoFile.createNewFile();
+                            }
+                            //selectedImageFile.mkdirs();
+
+                            FileOutputStream fOut = new FileOutputStream(capturedPhotoFile);
+
+                            capturedPhotoBitmap.compress(Bitmap.CompressFormat.PNG, 100, fOut);
+                            fOut.flush();
+                            fOut.close();
+
+                            crop(new File(getCacheDir(),
+                                    "temp"
+                            ).getPath());
+
+                    }
+                    return;
                 case PICK_IMAGE:
                     if (resultCode == RESULT_OK) {
                         Uri imageUri = data.getData();
@@ -137,7 +210,7 @@ public class CreateActivity extends BaseActivity {
                         Cursor cursor = getContentResolver().query(imageUri, new String[]{MediaStore.Images.ImageColumns.DATA}, null, null, null);
                         cursor.moveToFirst();
 
-                        final String imageFilePath = cursor.getString(0);
+                        final String imageFilePath = cursor.getString(cursor.getColumnIndex(MediaStore.Images.ImageColumns.DATA));
 
                         cursor.close();
 
@@ -146,26 +219,28 @@ public class CreateActivity extends BaseActivity {
                     break;
                 case PICK_PHOTO: {
 
-                    // todo cropper?
-                    if(resultCode == RESULT_OK) {
-                        Bundle extras = data.getExtras();
-                        Bitmap capturedPhotoBitmap = (Bitmap) extras.get("data");
-                        imageView.setScaleType(ImageView.ScaleType.CENTER_CROP);
-                        File capturedPhotoFile = new  File(getCacheDir(),
-                                "temp"
-                                );
-                        if(capturedPhotoFile.exists()){
-                            capturedPhotoFile.delete();
-                            capturedPhotoFile.createNewFile();
+                    if (resultCode == RESULT_OK) {
+                        if (data != null) {
+                            Bundle extras = data.getExtras();
+                            Bitmap capturedPhotoBitmap = (Bitmap) extras.get("data");
+                            File capturedPhotoFile = new File(getCacheDir(),
+                                    "temp"
+                            );
+                            if (capturedPhotoFile.exists()) {
+                                capturedPhotoFile.delete();
+                                capturedPhotoFile.createNewFile();
+                            }
+                            //selectedImageFile.mkdirs();
+
+                            FileOutputStream fOut = new FileOutputStream(capturedPhotoFile);
+
+                            capturedPhotoBitmap.compress(Bitmap.CompressFormat.PNG, 100, fOut);
+                            fOut.flush();
+                            fOut.close();
+                            crop(capturedPhotoFile.getPath());
+                        } else {
+                            crop(selectedImageFile.getPath());
                         }
-                        //selectedImageFile.mkdirs();
-
-                        FileOutputStream fOut = new FileOutputStream(capturedPhotoFile);
-
-                        capturedPhotoBitmap.compress(Bitmap.CompressFormat.PNG, 100, fOut);
-                        fOut.flush();
-                        fOut.close();
-                        crop(capturedPhotoFile.getPath());
                     }
                 }
                 break;
@@ -190,17 +265,17 @@ public class CreateActivity extends BaseActivity {
                         selectedCity = bundle.getString(LocationPickerActivity.EXTRA_CITY);
                         selectedLongitude = bundle.getDouble(LocationPickerActivity.EXTRA_LONGITUDE);
                         selectedLatitude = bundle.getDouble(LocationPickerActivity.EXTRA_LAT);
-                        if(selectedStreet==null || selectedStreet.equals("")){
+                        if (selectedStreet == null || selectedStreet.equals("")) {
                             pickLocationTextView.setText(R.string.create_pick_map_unknown);
                         } else {
-                            pickLocationTextView.setText(selectedStreet);
+                            pickLocationTextView.setText(selectedStreet + ", " + selectedCity);
                         }
                     }
                 }
                 break;
             }
 
-        }catch (Exception exp){
+        } catch (Exception exp) {
             bodyView.setText(exp.getLocalizedMessage());
         }
         super.onActivityResult(requestCode, resultCode, data);
@@ -211,7 +286,7 @@ public class CreateActivity extends BaseActivity {
 
         String filePath = imageFilePath;
         intent.putExtra(CropActivity.IMAGE_PATH, filePath);
-        intent.putExtra(CropActivity.IMAGE_PATH_RESULT, getCacheDir()+"/pic"+System.currentTimeMillis()+".jpg");
+        intent.putExtra(CropActivity.IMAGE_PATH_RESULT, getCacheDir() + "/pic" + System.currentTimeMillis() + ".jpg");
 
         startActivityForResult(intent, CROP_PICTURE);
 
@@ -239,40 +314,45 @@ public class CreateActivity extends BaseActivity {
                 return true;
             }
             case R.id.create:
-                if(selectedCategory==null){
-                    Toast.makeText(this, "Сначала выберите категорию", Toast.LENGTH_SHORT).show();
-                    return true;
-                }
-                final AlertDialog dialog = new AlertDialog.Builder(this)
-                        .setTitle("Отправка")
-                        .setMessage("Терпение")
-                        .setCancelable(false)
-                        .show();
-                String body = bodyView.getText().toString();
+                create();
 
-                CreateEventTask createEventTask = new CreateEventTask(
-                        body, selectedCategory.id,
-                        selectedLongitude,selectedLatitude,
-                        selectedCity, selectedStreet,
-                        selectedImageFile
-                ) {
-
-                    @Override
-                    protected void onSuccess(CreateEventResult result) {
-                        dialog.dismiss();
-                        finish();
-                    }
-
-                    @Override
-                    protected void onError(Exception exp) {
-                        dialog.dismiss();
-                        Toast.makeText(getBaseContext(), "Ошибка создания", Toast.LENGTH_SHORT).show();
-                    }
-                };
-                createEventTask.execute();
-                return true;
         }
         return super.onOptionsItemSelected(item);
+    }
+
+    private void create() {
+        if (selectedCategory == null) {
+            Toast.makeText(this, "Сначала выберите категорию", Toast.LENGTH_SHORT).show();
+            return;
+        }
+        final AlertDialog dialog = new AlertDialog.Builder(this)
+                .setTitle(R.string.create_sending)
+                .setMessage(R.string.create_sending_message)
+                .setCancelable(false)
+                .show();
+        String body = bodyView.getText().toString();
+
+        CreateEventTask createEventTask = new CreateEventTask(
+                body, selectedCategory.id,
+                selectedLongitude, selectedLatitude,
+                selectedCity, selectedStreet,
+                selectedImageFile
+        ) {
+
+            @Override
+            protected void onSuccess(CreateEventResult result) {
+                dialog.dismiss();
+                startActivity(EventActivity.getActivityIntent(getBaseContext(), result.event_id));
+                finish();
+            }
+
+            @Override
+            protected void onError(Exception exp) {
+                dialog.dismiss();
+                Toast.makeText(getBaseContext(), "Ошибка создания", Toast.LENGTH_SHORT).show();
+            }
+        };
+        createEventTask.execute();
     }
 
     public static Intent getActivityIntent(Context baseContext) {
